Renaming interaction.py -> mergedchoicetable.py
