+++ choicemodels/tools/__init__.py
@@ -1,4 +1,4 @@
 # ChoiceModels
 # See full license in LICENSE
 
-from .interaction import *
+from .mergedchoicetable import *
